Moving InitializationEvents to the agent
