+++ android-agent/build.gradle.kts
@@ -63,7 +63,7 @@ android {
 
 dependencies {
     implementation(project(":common"))
-    implementation(project(":instrumentation:startup")) // todo remove during the initializationevents adaptation work.
+    implementation(project(":instrumentation:common-api"))
     implementation(libs.androidx.core)
     implementation(libs.androidx.navigation.fragment)
     implementation(libs.androidx.lifecycle.process)
