+++ android-agent/src/main/java/io/opentelemetry/android/OpenTelemetryRumBuilder.java
@@ -14,9 +14,8 @@ import io.opentelemetry.android.config.OtelRumConfig;
 import io.opentelemetry.android.features.diskbuffering.DiskBufferingConfiguration;
 import io.opentelemetry.android.features.diskbuffering.SignalFromDiskExporter;
 import io.opentelemetry.android.features.diskbuffering.scheduler.ExportScheduleHandler;
+import io.opentelemetry.android.initialization.InitializationEvents;
 import io.opentelemetry.android.instrumentation.common.InstrumentedApplication;
-import io.opentelemetry.android.instrumentation.initialization.InitializationEvents;
-import io.opentelemetry.android.instrumentation.startup.SdkInitializationEvents;
 import io.opentelemetry.android.internal.features.networkattrs.NetworkAttributesSpanAppender;
 import io.opentelemetry.android.internal.features.persistence.DiskManager;
 import io.opentelemetry.android.internal.features.persistence.SimpleTemporaryFileProvider;
