+++ instrumentation/startup/build.gradle.kts
@@ -16,8 +16,8 @@ android {
 dependencies {
     api(platform(libs.opentelemetry.platform))
     api(libs.opentelemetry.api)
+    api(project(":android-agent"))
     api(project(":common"))
-    api(project(":instrumentation:common-api"))
     implementation(libs.androidx.core)
     implementation(libs.opentelemetry.semconv)
     implementation(libs.opentelemetry.sdk)
